requirements.txt
Flask~=3.0.3
Flask-Migrate~=4.0.7
Flask-Admin~=1.6.1
WTForms~=3.1.2
SQLAlchemy~=2.0.35
alembic~=1.13.3
python-dotenv~=1.0.1
Flask-Talisman~=1.1.0
Flask-Login~=0.6.3
cryptography~=43.0.3
pyotp~=2.9.0
MarkupSafe~=2.1.5
Flask-Limiter~=3.8.0